requirements.txt
Django~=5.0.6
python-dotenv~=1.0.1
djangorestframework~=3.15.2